Ultra Realistic 3D Ice cream Model
Using Blender, I have crafted an ultra realistic 3D model of an ice cream, meticulously designed to replicate the texture and appearance of a delectable frozen treat. This project showcases my expertise in 3D modeling, as I have incorporated high-quality textures and shaders to achieve a lifelike representation of the ice cream's surface, complete with intricate details such as melting drips and tantalizing toppings. The attention to detail and use of advanced techniques result in a visually stunning and mouth-watering virtual dessert, appealing to both the eyes and the taste buds.